[VM][Refactor] Move VM files to TVM runtime directory (#98)
