+++ include/tvm/relax/exec_builder.h
@@ -18,22 +18,21 @@
  */
 
 /*!
- * \file tvm/relax/vm/exec_builder.h
+ * \file tvm/relax/exec_builder.h
  */
-#ifndef TVM_RELAX_VM_EXEC_BUILDER_H_
-#define TVM_RELAX_VM_EXEC_BUILDER_H_
+#ifndef TVM_RELAX_EXEC_BUILDER_H_
+#define TVM_RELAX_EXEC_BUILDER_H_
 
 #include <tvm/ir/expr.h>
 #include <tvm/node/reflection.h>
 #include <tvm/node/repr_printer.h>
 #include <tvm/runtime/object.h>
 #include <tvm/runtime/registry.h>
+#include <tvm/runtime/relax_vm/bytecode.h>
+#include <tvm/runtime/relax_vm/executable.h>
 
 #include <string>
 #include <vector>
-
-#include "./bytecode.h"
-#include "./executable.h"
 
 namespace tvm {
 namespace relax {
@@ -121,4 +120,4 @@ class ExecBuilder : public ObjectRef {
 }  // namespace relax
 }  // namespace tvm
 
-#endif  // TVM_RELAX_VM_EXEC_BUILDER_H_
+#endif  // TVM_RELAX_EXEC_BUILDER_H_
